Bookmarking changes.
